move report components to report page pod
