+++ app/benchmarking/report/components/badge.tsx
