+++ app/benchmarking/report/components/mlcReportComponents/mlcAiSafetyRecipeRatingResult.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { RecipeGradeBadge } from '@/app/benchmarking/components/reportComponents/badge';
-import { gradeColorsMlc } from '@/app/benchmarking/components/reportComponents/gradeColors';
-import { gradingLettersMlcMap } from '@/app/benchmarking/components/reportComponents/mlcReportComponents/constants';
-import { RangedBarChart } from '@/app/benchmarking/components/reportComponents/rangedBarChart';
-import { RawRecipeMetricsScoresTable } from '@/app/benchmarking/components/reportComponents/rawScoresTable';
+import { RecipeGradeBadge } from '@/app/benchmarking/report/components/badge';
+import { gradeColorsMlc } from '@/app/benchmarking/report/components/gradeColors';
+import { gradingLettersMlcMap } from '@/app/benchmarking/report/components/mlcReportComponents/constants';
+import { RangedBarChart } from '@/app/benchmarking/report/components/rangedBarChart';
+import { RawRecipeMetricsScoresTable } from '@/app/benchmarking/report/components/rawScoresTable';
 import {
   GradingScale,
   RecipeEvaluationResult,
