+++ app/benchmarking/report/components/rangedBarChart.tsx
