+++ app/benchmarking/report/components/recipeRatingResult.tsx
