+++ app/benchmarking/report/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 import React from 'react';
-import { ReportViewer } from '@/app/benchmarking/components/reportComponents/reportViewer';
+import { ReportViewer } from '@/app/benchmarking/report/components/reportViewer';
 import { CookbookCategoryLabels } from '@/app/benchmarking/types/benchmarkReportTypes';
 import { CookbooksBenchmarkResult } from '@/app/benchmarking/types/benchmarkReportTypes';
 import { MainSectionSurface } from '@/app/components/mainSectionSurface';
